move the logic to a separate class
